From b105e7fe812da3ccaf7155c0fe14c8728b0d39a5 Mon Sep 17 00:00:00 2001 From: Mark Hatle <mark.hatle@windriver.com> Date: Mon, 20 Jan 2014 14:30:52 +0000 Subject: [PATCH] Add mechanism to attempt install without failing
In OpenEmbedded, for complementary and 'attemptonly' package processing,
we need a way to instruct smart to try to install, but ignore any
failures (usually conflicts).

This option only works for the install operation.

If a complementary install fails, an actual error occurred, one that
we can't ignore without losing the entire attempted transaction.  Keep
this as an error so that we can catch these cases in the futre.

Upstream-Status: Pending

This patch is backported from:
http://git.yoctoproject.org/cgit/cgit.cgi/poky/tree/meta/recipes-devtools/python/python-smartpm/smart-attempt.patch?id=yocto-2.0.1

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>
Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>
Signed-off-by: Lei Maohui <leimaohui@cn.fujitsu.com>